Fix sudo authentication
Added shadow package for sudo authentication. Sudo tried to use shadow for authentication but the package was not included in the image so password login for sudo failed. The beagle user's hash in /etc/shadow/ had the prefix $5$ indicating that it was hashed with SHA256, that is why SHADOW_SHA_CRYPT was chosen.