Add led commands (ledcmd)
This adds commands for on|off|heartbeat|blink. The imlementation uses LED sysfs. The linux kernel did not have the necessary triggers, so a linux-fragment.config has been added to add the missing triggers. For non-root users (beagle user in this case) it was annoying to have to type password for each command. Therefore a sudoers.d file was added to allow the beagle user to run ledcmd without having to enter password. Sudo is still necessary though.